requirements.txt
virtualenv
django==1.11
djangorestframework
requests
pillow
scrapy
mongoengine
pymongo
schema
git+https://github.com/django-nonrel/django@nonrel-1.5
git+https://github.com/django-nonrel/djangotoolbox
git+https://github.com/django-nonrel/mongodb-engine
scrapyd
python-scrapyd-api
scrapy-splash